# add goja as JS VM for transformers