--- a/adaptor/transformer/gojajs/testdata/transformer.js
+++ b/adaptor/transformer/gojajs/testdata/transformer.js
@@ -1,0 +1,1 @@
+function transform(doc) { return doc }