--- a/adaptor/transformer/transformer_test.go
+++ b/adaptor/transformer/transformer_test.go
@@ -22,6 +22,7 @@ func TestSampleConfig(t *testing.T) {
 
 var initTests = []map[string]interface{}{
 	{"filename": "testdata/transformer.js"},
+	{"filename": "testdata/goja-transformer.js", "vm": "goja"},
 }
 
 func TestInit(t *testing.T) {