--- a/glide.lock
+++ b/glide.lock
@@ -5,6 +5,17 @@ imports:
   version: b02f2bbce11d7ea6b97f282ef1771b0fe2f65ef3
 - name: github.com/compose/mejson
   version: afcf51c7c640b1e09c17065d4e5b8de3259d6261
+- name: github.com/dlclark/regexp2
+  version: 3f97b6a39134e6bd88a72d1388d1632e60d83ef6
+  subpackages:
+  - syntax
+- name: github.com/dop251/goja
+  version: 1b1607cf5cf9c00ae9c2b0da9ac4f48cb25f8703
+  subpackages:
+  - ast
+  - file
+  - parser
+  - token
 - name: github.com/golang/protobuf
   version: 4bd1920723d7b7c925de087aa32e2187708897f7
   subpackages:
@@ -54,6 +65,17 @@ imports:
   version: d75a52659825e75fff6158388dddc6a5b04f9ba5
   subpackages:
   - unix
+- name: golang.org/x/text
+  version: f28f36722d5ef2f9655ad3de1f248e3e52ad5ebd
+  subpackages:
+  - cases
+  - collate
+  - internal
+  - internal/colltab
+  - internal/tag
+  - language
+  - transform
+  - unicode/norm
 - name: gopkg.in/fatih/pool.v2
   version: 20a0a429c5f93de45c90f5f09ea297c25e0929b3
 - name: gopkg.in/gorethink/gorethink.v3