[goja](github.com/dop251/goja) looks to be a nice alternative to offer people who may need a little more performance from the transformer function. Initial benchmarks showed the following results:

```shell
$ benchcmp otto.profile gojajs/gojajs.profile
benchmark                   old ns/op     new ns/op     delta
BenchmarkTransformOne-8     13565         6265          -53.81%
```

- [x] CHANGELOG.md updated
- [x] README.md updated

fixes #230